added 'step-through-debugger'
Wrote and head-bashed a ridiculous bug: I was generating
the rectangles in a DEPTH FIRST manner, first for depth 0,
then for depth 1, etc.

What I needed was BFS, so that, as we descend each layer,
the boxes are made in the same order (and thus receive the
same RNG state).